refactor: extract HTTP utils to companion module (CDMCH-55)
Move 8 helper functions and sensitive-data constants from client.ts
to httpUtils.ts. Original module re-exports for backward compatibility.
